22.0.14-preview

+++ crates/libs/sys/src/Windows/Win32/System/SystemServices/mod.rs
@@ -3447,7 +3447,7 @@ impl ::core::clone::Clone for IMAGE_DOS_HEADER {
     }
 }
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
-pub const IMAGE_DOS_SIGNATURE: u32 = 23117u32;
+pub const IMAGE_DOS_SIGNATURE: u16 = 23117u16;
 #[repr(C, packed(1))]
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
 pub struct IMAGE_DYNAMIC_RELOCATION32 {
@@ -3789,7 +3789,7 @@ impl ::core::clone::Clone for IMAGE_LINENUMBER_0 {
     }
 }
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
-pub const IMAGE_NT_SIGNATURE: u32 = 17744u32;
+pub const IMAGE_NT_SIGNATURE: u16 = 17744u16;
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
 pub const IMAGE_NUMBEROF_DIRECTORY_ENTRIES: u32 = 16u32;
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
@@ -3842,9 +3842,9 @@ impl ::core::clone::Clone for IMAGE_OS2_HEADER {
     }
 }
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
-pub const IMAGE_OS2_SIGNATURE: u32 = 17742u32;
+pub const IMAGE_OS2_SIGNATURE: u16 = 17742u16;
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
-pub const IMAGE_OS2_SIGNATURE_LE: u32 = 17740u32;
+pub const IMAGE_OS2_SIGNATURE_LE: u16 = 17740u16;
 #[repr(C)]
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`, `\"Win32_Foundation\"`*"]
 #[cfg(feature = "Win32_Foundation")]
@@ -5013,7 +5013,7 @@ impl ::core::clone::Clone for IMAGE_VXD_HEADER {
     }
 }
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
-pub const IMAGE_VXD_SIGNATURE: u32 = 17740u32;
+pub const IMAGE_VXD_SIGNATURE: u16 = 17740u16;
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
 pub const IMAGE_WEAK_EXTERN_ANTI_DEPENDENCY: u32 = 4u32;
 #[doc = "*Required features: `\"Win32_System_SystemServices\"`*"]
